Webpack 4 configuration
