+++ package.json
@@ -57,6 +57,7 @@
     "react-router-dom": "^4.3.1",
     "sass-loader": "^7.0.3",
     "style-loader": "^0.21.0",
+    "uglifyjs-webpack-plugin": "^1.2.7",
     "url-loader": "^1.0.1",
     "webpack": "^4.16.3",
     "webpack-bundle-analyzer": "^2.13.1",
